Renaming all "driftory-viewer" references to "driftory"
