+++ src/demo/demo.js
@@ -1,10 +1,10 @@
-import DriftoryViewer from '../library/driftory-viewer';
+import Driftory from '../library/driftory';
 
 var driftoryViewer, nextButton, previousButton;
 
 // ----------
 function init() {
-  var driftoryViewer = new DriftoryViewer({
+  var driftory = new Driftory({
     container: document.querySelector('.driftory-viewer-container')
   });
 
@@ -12,15 +12,15 @@ function init() {
   previousButton = document.querySelector('.previous-button');
 
   nextButton.addEventListener('click', function () {
-    var index = driftoryViewer.getFrameIndex();
-    index = (index + 1) % driftoryViewer.getFrameCount();
-    driftoryViewer.goToFrame(index);
+    var index = driftory.getFrameIndex();
+    index = (index + 1) % driftory.getFrameCount();
+    driftory.goToFrame(index);
   });
 
   previousButton.addEventListener('click', function () {
-    var index = driftoryViewer.getFrameIndex();
-    index = index === 0 ? driftoryViewer.getFrameCount() - 1 : index - 1;
-    driftoryViewer.goToFrame(index);
+    var index = driftory.getFrameIndex();
+    index = index === 0 ? driftory.getFrameCount() - 1 : index - 1;
+    driftory.goToFrame(index);
   });
 
   fetch('comic.json')
@@ -34,7 +34,7 @@ function init() {
     .then(function (json) {
       // console.log(json);
 
-      driftoryViewer.openComic(json.comic);
+      driftory.openComic(json.comic);
     });
 }
